refactor(client-cli): move the Cardano node version predicates to the version module
They were duplicated in the snapshot converter and the Cardano database shared
steps.
